Refactor ui code in multiple files
